--- src/field/richtext/InsertMenu.module.css
@@ -1,19 +1,0 @@
-.InsertMenu-trigger {
-  display: inline-flex;
-  align-items: center;
-  gap: 8px;
-  border-radius: 8px;
-  background: var(--alinea-muted);
-  border: 1px solid var(--alinea-border);
-}
-
-.InsertMenu-menu {
-  z-index: 1;
-  max-width: none;
-}
-
-.InsertMenu-item {
-  display: inline-flex;
-  align-items: center;
-  gap: 8px;
-}
