+++ src/field/richtext/RichTextField.module.css
@@ -6,44 +6,66 @@
   outline: 3px solid transparent;
   transition: var(--alinea-transition);
   transition-property: border-color, outline-color, background-color;
+}
 
-  &:hover {
-    border-color: var(--alinea-border-hover);
-    outline-color: var(--alinea-ring-subtle);
-  }
+.RichTextFieldView:hover {
+  border-color: var(--alinea-border-hover);
+  outline-color: var(--alinea-ring-subtle);
+}
 
-  &:focus-within {
-    border-color: var(--alinea-border-accent);
-    outline-color: var(--alinea-primary-overlay);
-  }
+.RichTextFieldView:focus-within {
+  border-color: var(--alinea-border-accent);
+  outline-color: var(--alinea-primary-overlay);
+}
 
-  &[data-invalid='true'] {
-    border-color: var(--alinea-border-danger);
-    outline-color: var(--alinea-danger-overlay);
-  }
+.RichTextFieldView[data-focused='true'] {
+  border-color: var(--alinea-border-accent);
+  outline-color: var(--alinea-primary-overlay);
+}
 
-  &[data-read-only='true'] {
-    background-color: var(--alinea-input-disabled);
-  }
+.RichTextFieldView:focus-within:not([data-focused='true']) {
+  border-color: var(--alinea-border);
+  outline-color: transparent;
+}
 
-  &[data-read-only='true']:hover,
-  &[data-read-only='true']:focus-within {
-    border-color: var(--alinea-border);
-    outline-color: transparent;
-  }
+.RichTextFieldView[data-invalid='true'] {
+  border-color: var(--alinea-border-danger);
+  outline-color: var(--alinea-danger-overlay);
+}
 
-  :global(.ProseMirror) {
-    outline: none;
-    min-height: calc(1.5em + 16px);
-    padding: 8px 10px;
-    color: var(--alinea-fg);
-    font-size: var(--alinea-font-size-base);
-    line-height: 1.5;
-  }
+.RichTextFieldView[data-read-only='true'] {
+  background-color: var(--alinea-input-disabled);
+}
 
-  &[data-read-only='true'] :global(.ProseMirror) {
-    color: var(--alinea-fg-disabled);
-  }
+.RichTextFieldView[data-read-only='true']:hover,
+.RichTextFieldView[data-read-only='true']:focus-within {
+  border-color: var(--alinea-border);
+  outline-color: transparent;
+}
+
+.RichTextFieldView :global(.ProseMirror) {
+  min-height: calc(1.5em + 16px);
+  padding: 8px 10px;
+  outline: none;
+  color: var(--alinea-fg);
+  font-size: var(--alinea-font-size-base);
+  line-height: 1.5;
+}
+
+.RichTextFieldView[data-read-only='true'] :global(.ProseMirror) {
+  color: var(--alinea-fg-disabled);
+}
+
+.RichTextFieldView[data-inline='true'] :global(.ProseMirror) {
+  min-height: auto;
+}
+
+.RichTextFieldView-unsupportedBlock {
+  margin: 6px 10px;
+  padding: 8px 10px;
+  color: var(--alinea-fg-muted);
+  border: 1px dashed var(--alinea-border);
+  border-radius: 6px;
 }
 
 .RichTextFieldCompactView {
@@ -61,9 +83,4 @@
 
 .RichTextFieldCompactView[data-empty] {
   color: var(--alinea-fg-muted);
-}
-
-.RichTextFieldBlock {
-  margin: 14px 0;
-  outline: 3px solid transparent;
 }
