+++ src/field/richtext/RichTextToolbar.module.css
@@ -1,13 +1,13 @@
 .RichTextToolbar {
+  display: flex;
   width: 100%;
   height: 40px;
   min-width: 0;
-  display: flex;
   align-items: center;
+  box-sizing: border-box;
   overflow-x: auto;
   overflow-y: hidden;
-  box-sizing: border-box;
-  scrollbar-width: none;
   padding: 4px 12px;
   background: var(--alinea-bg);
+  scrollbar-width: none;
 }
