+++ src/field/richtext/RichTextToolbar.tsx
@@ -10,299 +10,268 @@ import {
 } from '#/components.js'
 import type {Reference} from '#/core/Reference.js'
 import {entries} from '#/core/util/Objects.js'
-import type {
-  PickTextLinkFunc,
-  PickerValue
-} from '#/field/richtext/PickTextLink.js'
-import {
-  attributesToReference,
-  referenceToAttributes
-} from '#/field/richtext/ReferenceLink.js'
-import type {
-  RichTextToolbarContext,
-  ToolbarButton,
-  ToolbarConfig,
-  ToolbarGroup as ToolbarConfigGroup,
-  ToolbarMenu
-} from '#/field/richtext/Toolbar.js'
 import type {UrlReference} from '#/picker/url.js'
 import styler from '@alinea/styler'
 import type {Editor} from '@tiptap/core'
+import {useEditorState} from '@tiptap/react'
 import {useMemo, type ReactNode} from 'react'
+import type {PickTextLinkFunc, PickerValue} from './PickTextLink.js'
+import {attributesToReference, referenceToAttributes} from './ReferenceLink.js'
+import {
+  defaultToolbar,
+  type RichTextToolbarContext,
+  type ToolbarButton,
+  type ToolbarConfig,
+  type ToolbarGroup as ToolbarConfigGroup,
+  type ToolbarMenu
+} from './Toolbar.js'
 import css from './RichTextToolbar.module.css'
-import {defaultToolbar} from './Toolbar.js'
 
 const styles = styler(css)
-
-export type RichTextCommand = () => ReturnType<Editor['chain']>
 
 export interface RichTextToolbarProps {
   editor: Editor
   enableTables?: boolean
-  focusToggle: (target: EventTarget | null) => void
+  ownerId: string
   pickLink?: PickTextLinkFunc
   toolbar?: ToolbarConfig
-}
-
-export function createToolbarExec(editor: Editor): RichTextCommand {
-  return () => editor.chain().focus(null, {scrollIntoView: false})
-}
-
-export function createLinkHandler(
-  editor: Editor,
-  pickLink: PickTextLinkFunc | undefined,
-  exec = createToolbarExec(editor)
-) {
-  return function handleLink() {
-    const attrs = editor.getAttributes('link')
-    const {view} = editor
-    const {from, to} = view.state.selection
-    const isSelection = from !== to
-    if (!pickLink) return handleBrowserLink(editor, exec, attrs, isSelection)
-    const existing: Reference | undefined = attributesToReference(attrs)
-    return pickLink({
-      link: existing,
-      title: attrs.title,
-      blank: attrs.target === '_blank',
-      hasLink: Boolean(existing),
-      requireDescription: !isSelection
-    })
-      .then(picked => {
-        if (picked === undefined) return
-        if (!picked.link) {
-          exec().unsetLink().run()
-          return
-        }
-        const linkAttributes = createLinkAttributes(picked)
-        if (existing) {
-          exec().extendMarkRange('link').setLink(linkAttributes).run()
-        } else if (isSelection) {
-          exec().setLink(linkAttributes).run()
-        } else {
-          exec()
-            .insertContent({
-              type: 'text',
-              text:
-                picked.description ||
-                (picked.link as UrlReference)._title ||
-                (picked.link as UrlReference)._url ||
-                '',
-              marks: [{type: 'link', attrs: linkAttributes}]
-            })
-            .run()
-        }
-      })
-      .catch(console.error)
-  }
+  onFocusChange: (focused: boolean, nextTarget?: EventTarget | null) => void
 }
 
 export function RichTextToolbar({
   editor,
   enableTables,
-  focusToggle,
+  ownerId,
   pickLink,
-  toolbar
+  toolbar,
+  onFocusChange
 }: RichTextToolbarProps) {
+  useEditorState({editor, selector: context => context.transactionNumber})
   const config = useMemo(
-    () => toolbar ?? defaultToolbar(enableTables || false),
+    () => toolbar ?? defaultToolbar(Boolean(enableTables)),
     [enableTables, toolbar]
   )
-  const ctx = useMemo(() => {
-    const exec = createToolbarExec(editor)
+  const context = useMemo(() => {
+    const exec = () => editor.chain().focus(null, {scrollIntoView: false})
     return {
       editor,
-      focusToggle,
-      pickLink: pickLink ?? pickBrowserLink,
       enableTables,
       exec,
+      focusToggle: target => onFocusChange(Boolean(target)),
+      pickLink: pickLink ?? emptyPicker,
       handleLink: createLinkHandler(editor, pickLink, exec),
       toolbar: config
     } satisfies RichTextToolbarContext
-  }, [config, editor, enableTables, focusToggle, pickLink])
+  }, [config, editor, enableTables, onFocusChange, pickLink])
+
   return (
     <div
-      tabIndex={-1}
       className={styles.RichTextToolbar()}
       data-richtext-toolbar="true"
-      onFocus={event => ctx.focusToggle(event.currentTarget)}
-      onBlur={event => ctx.focusToggle(event.relatedTarget)}
+      data-richtext-toolbar-owner={ownerId}
+      onFocus={() => onFocusChange(true)}
+      onBlur={event => onFocusChange(false, event.relatedTarget)}
     >
-      <Toolbar aria-label="Text formatting" data-orientation="horizontal">
-        <ToolbarItems config={config} ctx={ctx} subMenu={false} />
+      <Toolbar aria-label="Text formatting">
+        <ToolbarItems
+          config={config}
+          context={context}
+          menu={false}
+          ownerId={ownerId}
+        />
       </Toolbar>
     </div>
   )
 }
 
-interface ToolbarButtonProps {
-  button: ToolbarButton
-  ctx: RichTextToolbarContext
-}
-
-function ToolbarButtonView({button, ctx}: ToolbarButtonProps) {
-  const icon = button.icon?.(ctx)
-  const label = resolveNode(button.label, ctx)
-  const active = button.active?.(ctx)
-  const title = button.title ?? (typeof label === 'string' ? label : undefined)
-  return (
-    <Button
-      size="icon-nav"
-      appearance={active ? 'active' : 'plain'}
-      isDisabled={button.disabled?.(ctx)}
-      onPress={() => button.onSelect(ctx)}
-      aria-label={title}
-    >
-      {icon && <Icon icon={icon} data-slot="icon" />}
-      {!icon && label}
-    </Button>
-  )
-}
-
-interface ToolbarMenuProps {
-  ctx: RichTextToolbarContext
-  menu: ToolbarMenu
-}
-
-function ToolbarMenuView({ctx, menu}: ToolbarMenuProps) {
-  const icon = menu.icon?.(ctx)
-  const label = resolveNode(menu.label, ctx)
-  const config = typeof menu.items === 'function' ? menu.items(ctx) : menu.items
-  const textValue = getTextValue(label) ?? 'Menu'
-  return (
-    <Menu
-      aria-label={textValue}
-      label={
-        <Button appearance="plain">
-          {icon && <Icon icon={icon} data-slot="icon" />}
-          {label}
-        </Button>
-      }
-    >
-      <ToolbarItems config={config} ctx={ctx} subMenu={true} />
-    </Menu>
-  )
-}
-
 interface ToolbarItemsProps {
   config: ToolbarConfig
-  ctx: RichTextToolbarContext
-  subMenu: boolean
+  context: RichTextToolbarContext
+  menu: boolean
+  ownerId: string
 }
 
-function ToolbarItems({config, ctx, subMenu}: ToolbarItemsProps) {
-  return entries(config).reduce((result, [name, entry], index) => {
-    if (index > 0) {
-      result.push(
-        subMenu ? (
-          <MenuSeparator key={`${name}-separator`} />
-        ) : (
-          <ToolbarSeparator key={`${name}-separator`} />
-        )
-      )
-    }
-    result.push(renderEntry(name, entry, ctx, subMenu))
-    return result
-  }, Array<ReactNode>())
+function ToolbarItems({config, context, menu, ownerId}: ToolbarItemsProps) {
+  return entries(config).flatMap(([name, entry], index) => {
+    const separator =
+      index === 0
+        ? []
+        : [
+            menu ? (
+              <MenuSeparator key={`${name}-separator`} />
+            ) : (
+              <ToolbarSeparator key={`${name}-separator`} />
+            )
+          ]
+    return [...separator, renderEntry(name, entry, context, menu, ownerId)]
+  })
 }
 
 function renderEntry(
   name: string,
   entry: ToolbarButton | ToolbarMenu | ToolbarConfigGroup,
-  ctx: RichTextToolbarContext,
-  subMenu: boolean
+  context: RichTextToolbarContext,
+  menu: boolean,
+  ownerId: string
 ): ReactNode {
   if ('items' in entry) {
-    return <ToolbarMenuView key={name} menu={entry} ctx={ctx} />
+    const label = resolve(entry.label, context)
+    const items =
+      typeof entry.items === 'function' ? entry.items(context) : entry.items
+    return (
+      <Menu
+        key={name}
+        aria-label={text(label) ?? humanize(name)}
+        popoverProps={{'data-richtext-toolbar-owner': ownerId}}
+        label={
+          <Button appearance="plain">
+            {entry.icon && <Icon icon={entry.icon(context)} data-slot="icon" />}
+            {label}
+          </Button>
+        }
+      >
+        <ToolbarItems config={items} context={context} menu ownerId={ownerId} />
+      </Menu>
+    )
   }
   if ('group' in entry) {
-    const config =
-      typeof entry.group === 'function' ? entry.group(ctx) : entry.group
-    const content: Array<ReactNode> = entries(config).map(
-      ([groupName, groupEntry]) =>
-        renderEntry(`${name}-${groupName}`, groupEntry, ctx, subMenu)
+    const group =
+      typeof entry.group === 'function' ? entry.group(context) : entry.group
+    const children = entries(group).map(([childName, child]) =>
+      renderEntry(`${name}-${childName}`, child, context, menu, ownerId)
     )
-    if (subMenu) return content
-    return <ToolbarGroup key={name}>{content}</ToolbarGroup>
+    return menu ? children : <ToolbarGroup key={name}>{children}</ToolbarGroup>
   }
-  if (subMenu) {
-    const label = resolveNode(entry.label, ctx)
-    const title =
-      entry.title ?? getTextValue(label) ?? humanizeToolbarName(name)
-    const icon = entry.icon?.(ctx)
+  const label = resolve(entry.label, context)
+  const title = entry.title ?? text(label) ?? humanize(name)
+  if (menu) {
     return (
       <MenuItem
         key={name}
         id={name}
         textValue={title}
-        isDisabled={entry.disabled?.(ctx)}
-        onAction={() => entry.onSelect(ctx)}
+        isDisabled={entry.disabled?.(context)}
+        onAction={() => entry.onSelect(context)}
       >
-        {icon && <Icon icon={icon} data-slot="icon" />}
+        {entry.icon && <Icon icon={entry.icon(context)} data-slot="icon" />}
         <span>{label ?? title}</span>
       </MenuItem>
     )
   }
-  return <ToolbarButtonView key={name} button={entry} ctx={ctx} />
+  return (
+    <Button
+      key={name}
+      size="icon-nav"
+      appearance={entry.active?.(context) ? 'active' : 'plain'}
+      aria-label={title}
+      isDisabled={entry.disabled?.(context)}
+      onPress={() => entry.onSelect(context)}
+    >
+      {entry.icon ? (
+        <Icon icon={entry.icon(context)} data-slot="icon" />
+      ) : (
+        label
+      )}
+    </Button>
+  )
 }
 
-function handleBrowserLink(
+function createLinkHandler(
   editor: Editor,
-  exec: RichTextCommand,
-  attrs: Record<string, unknown>,
-  isSelection: boolean
+  picker: PickTextLinkFunc | undefined,
+  exec: () => ReturnType<Editor['chain']>
 ) {
-  if (typeof window === 'undefined') return
-  const currentHref = typeof attrs.href === 'string' ? attrs.href : ''
-  const href = window.prompt('Enter link URL', currentHref)
-  if (href === null) return
-  const nextHref = href.trim()
-  if (!nextHref) {
-    exec().unsetLink().run()
-    return
+  return function handleLink() {
+    const attributes = editor.getAttributes('link')
+    const {from, to} = editor.state.selection
+    const selected = from !== to
+    if (!picker) return browserLink(editor, exec, attributes, selected)
+    const existing: Reference | undefined = attributesToReference(attributes)
+    void picker({
+      link: existing,
+      title: attributes.title,
+      blank: attributes.target === '_blank',
+      hasLink: Boolean(existing),
+      requireDescription: !selected
+    }).then(picked => applyLink(editor, exec, picked, existing, selected))
   }
-  const linkAttributes = {href: nextHref}
-  if (editor.isActive('link')) {
-    exec().extendMarkRange('link').setLink(linkAttributes).run()
-  } else if (isSelection) {
-    exec().setLink(linkAttributes).run()
-  } else {
+}
+
+function applyLink(
+  editor: Editor,
+  exec: () => ReturnType<Editor['chain']>,
+  picked: PickerValue | undefined,
+  existing: Reference | undefined,
+  selected: boolean
+) {
+  if (picked === undefined) return
+  if (!picked.link) return void exec().unsetLink().run()
+  const attributes = {
+    title: picked.title,
+    ...referenceToAttributes(picked.link),
+    target:
+      (picked.link as UrlReference)._target ??
+      (picked.blank ? '_blank' : undefined)
+  }
+  if (existing) exec().extendMarkRange('link').setLink(attributes).run()
+  else if (selected) exec().setLink(attributes).run()
+  else
     exec()
       .insertContent({
         type: 'text',
-        text: nextHref,
-        marks: [{type: 'link', attrs: linkAttributes}]
+        text:
+          picked.description ||
+          (picked.link as UrlReference)._title ||
+          (picked.link as UrlReference)._url ||
+          '',
+        marks: [{type: 'link', attrs: attributes}]
       })
       .run()
-  }
 }
 
-function createLinkAttributes(picked: PickerValue) {
-  const link = picked.link!
-  return {
-    title: picked.title,
-    ...referenceToAttributes(link),
-    target:
-      (link as UrlReference)._target ?? (picked.blank ? '_blank' : undefined)
-  }
-}
-
-function resolveNode(
-  value: ToolbarButton['label'] | ToolbarMenu['label'],
-  ctx: RichTextToolbarContext
+function browserLink(
+  editor: Editor,
+  exec: () => ReturnType<Editor['chain']>,
+  attributes: Record<string, unknown>,
+  selected: boolean
 ) {
-  return typeof value === 'function' ? value(ctx) : value
+  const href = window.prompt(
+    'Enter link URL',
+    typeof attributes.href === 'string' ? attributes.href : ''
+  )
+  if (href === null) return
+  const value = href.trim()
+  if (!value) return void exec().unsetLink().run()
+  if (editor.isActive('link'))
+    exec().extendMarkRange('link').setLink({href: value}).run()
+  else if (selected) exec().setLink({href: value}).run()
+  else
+    exec()
+      .insertContent({
+        type: 'text',
+        text: value,
+        marks: [{type: 'link', attrs: {href: value}}]
+      })
+      .run()
 }
 
-function getTextValue(value: ReactNode) {
+function resolve(
+  value: ToolbarButton['label'] | ToolbarMenu['label'],
+  context: RichTextToolbarContext
+) {
+  return typeof value === 'function' ? value(context) : value
+}
+
+function text(value: ReactNode): string | undefined {
   return typeof value === 'string' ? value : undefined
 }
 
-function humanizeToolbarName(name: string) {
-  return name
+function humanize(value: string): string {
+  return value
     .replace(/([a-z0-9])([A-Z])/g, '$1 $2')
     .replace(/(^|\s)\w/g, letter => letter.toUpperCase())
 }
 
-async function pickBrowserLink() {
+async function emptyPicker(): Promise<undefined> {
   return undefined
 }
