--- src/field/richtext2/RichTextField.module.css
@@ -1,86 +1,0 @@
-.RichTextFieldView {
-  overflow: hidden;
-  background-color: var(--alinea-input);
-  border: 1px solid var(--alinea-border);
-  border-radius: 8px;
-  outline: 3px solid transparent;
-  transition: var(--alinea-transition);
-  transition-property: border-color, outline-color, background-color;
-}
-
-.RichTextFieldView:hover {
-  border-color: var(--alinea-border-hover);
-  outline-color: var(--alinea-ring-subtle);
-}
-
-.RichTextFieldView:focus-within {
-  border-color: var(--alinea-border-accent);
-  outline-color: var(--alinea-primary-overlay);
-}
-
-.RichTextFieldView[data-focused='true'] {
-  border-color: var(--alinea-border-accent);
-  outline-color: var(--alinea-primary-overlay);
-}
-
-.RichTextFieldView:focus-within:not([data-focused='true']) {
-  border-color: var(--alinea-border);
-  outline-color: transparent;
-}
-
-.RichTextFieldView[data-invalid='true'] {
-  border-color: var(--alinea-border-danger);
-  outline-color: var(--alinea-danger-overlay);
-}
-
-.RichTextFieldView[data-read-only='true'] {
-  background-color: var(--alinea-input-disabled);
-}
-
-.RichTextFieldView[data-read-only='true']:hover,
-.RichTextFieldView[data-read-only='true']:focus-within {
-  border-color: var(--alinea-border);
-  outline-color: transparent;
-}
-
-.RichTextFieldView :global(.ProseMirror) {
-  min-height: calc(1.5em + 16px);
-  padding: 8px 10px;
-  outline: none;
-  color: var(--alinea-fg);
-  font-size: var(--alinea-font-size-base);
-  line-height: 1.5;
-}
-
-.RichTextFieldView[data-read-only='true'] :global(.ProseMirror) {
-  color: var(--alinea-fg-disabled);
-}
-
-.RichTextFieldView[data-inline='true'] :global(.ProseMirror) {
-  min-height: auto;
-}
-
-.RichTextFieldView-unsupportedBlock {
-  margin: 6px 10px;
-  padding: 8px 10px;
-  color: var(--alinea-fg-muted);
-  border: 1px dashed var(--alinea-border);
-  border-radius: 6px;
-}
-
-.RichTextFieldCompactView {
-  display: inline-block;
-  min-width: 0;
-  max-width: 100%;
-  overflow: hidden;
-  color: var(--alinea-fg);
-  font-size: var(--alinea-font-size-base);
-  line-height: 1.3;
-  text-overflow: ellipsis;
-  vertical-align: middle;
-  white-space: nowrap;
-}
-
-.RichTextFieldCompactView[data-empty] {
-  color: var(--alinea-fg-muted);
-}
