--- src/field/richtext2/RichTextToolbar.module.css
@@ -1,13 +1,0 @@
-.RichTextToolbar {
-  display: flex;
-  width: 100%;
-  height: 40px;
-  min-width: 0;
-  align-items: center;
-  box-sizing: border-box;
-  overflow-x: auto;
-  overflow-y: hidden;
-  padding: 4px 12px;
-  background: var(--alinea-bg);
-  scrollbar-width: none;
-}
